refactor(uttaksplan): slå saman feltSynlighet og useFeltSynlighet
Flyttar useFeltSynlighet-hooken inn i feltSynlighet.ts og slettar
den separate fila for å halda alt samla.
