+++ packages/uttaksplan/src/regler/synlighet/index.ts
@@ -4,9 +4,8 @@ import { FORELDER_VALG_KAPITTEL } from './forelderValg';
 
 export { synlighetForForelderValg } from './forelderValg';
 export type { ForelderValgKontekst } from './forelderValg';
-export { synlighetForFelter } from './feltSynlighet';
+export { synlighetForFelter, useFeltSynlighet } from './feltSynlighet';
 export type { FeltSynlighetKontekst } from './feltSynlighet';
-export { useFeltSynlighet } from './useFeltSynlighet';
 
 /**
  * Alle synlighetsregler i uttaksplan. Brukes av Storybook-siden for å
